+++ ruoyi-modules/xanadu-distribution-center/src/main/java/edu/neu/dbc/controller/SupplierController.java
@@ -2,10 +2,13 @@ package edu.neu.dbc.controller;
 
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.ruoyi.common.core.web.domain.AjaxResult;
 import edu.neu.dbc.entity.Product;
 import edu.neu.dbc.entity.PurchaseRecord;
 import edu.neu.dbc.entity.Supplier;
+import edu.neu.dbc.service.ProductService;
 import edu.neu.dbc.service.PurchaseRecordService;
 import edu.neu.dbc.service.SupplierService;
 import io.swagger.annotations.ApiOperation;
@@ -40,11 +43,13 @@ public class SupplierController {
 
     @Autowired
     PurchaseRecordService purchaseRecordService;
+    @Autowired
+    ProductService productService;
 
 
-    @GetMapping("/list")
+    @GetMapping("/listAll")
     @ApiOperation("查询所有供应商")
-    public AjaxResult list() {
+    public AjaxResult listAll() {
         return AjaxResult.success(supplierService.list());
     }
 
@@ -57,6 +62,20 @@ public class SupplierController {
             return AjaxResult.error("供应商不存在");
         }
         return AjaxResult.success(supplier);
+    }
+
+    @GetMapping("/list/{pageNum}/{pageSize}")
+    @ApiOperation("分页查询供应商")
+    public AjaxResult pageList(@PathVariable Long pageNum, @PathVariable Long pageSize) {
+        return AjaxResult.success(supplierService.page(new Page<>(pageNum,pageSize)));
+    }
+
+    @GetMapping("/query/{pageNum}/{pageSize}")
+    @ApiOperation("根据姓名、地址查询供应商")
+    public AjaxResult query(@PathVariable Long pageNum, @PathVariable Long pageSize,@RequestParam(value = "name" ,required=false) String name,@RequestParam(value = "address" ,required=false) String address) {
+        QueryWrapper queryWrapper=new QueryWrapper();
+        queryWrapper.like(name!=null,"name",name).like(address!=null,"address",address);
+        return AjaxResult.success(supplierService.page(new Page<>(pageNum,pageSize),queryWrapper));
     }
 
     @PostMapping("/add")
@@ -84,6 +103,11 @@ public class SupplierController {
         QueryWrapper<PurchaseRecord> purchaseRecordQueryWrapper = new QueryWrapper<PurchaseRecord>().eq("supplier_id", id);
         if (purchaseRecordService.count(purchaseRecordQueryWrapper) > 0) {
             return AjaxResult.error("该供应商有过采购记录，无法删除");
+        }
+        //还需要找到是否存在商品由该供货商供货，如果存在，也不能删除
+        QueryWrapper<Product> productQueryWrapper = new QueryWrapper<Product>().eq("supplier_id", id);
+        if (productService.count(productQueryWrapper) > 0) {
+            return AjaxResult.error("该供应商有过商品供货，无法删除");
         }
         return AjaxResult.success(supplierService.removeById(id));
     }
